Seperate components in detail page
